ULMUD-019/playerhealth
Why?
Players need a health variable in order to implement combat into the
game.

How?
Added a health variable to the player class, ensuring that every new
created character starts with a base health value of 20. Also added a
"check health" command to see your current health v alue, along with a
test screenplay for this. Updated the default players in players.txt.