ULMUD-022/monsterattack
Why?
Monsters need to be able to attack players in order for combat to be
difficult.

How?
Added a "behavior" variable to monsters that will make a monster hostile
and unable to move once it is attacked. It will attack any players in the
same room back until it is no longer hostile, or once the player leaves
the room.